+++ tools/benchmarks/monotonic_pq.cpp
@@ -270,7 +270,7 @@ int main(int argc, char* argv[])
     {
         if ((i % mega) == 0)
             LOG1 << std::fixed << std::setprecision(2) << std::setw(5)
-                 << (100.0 * (double)i / (double)nelements) << "% "
+                 << (100.0 * static_cast<double>(i) / nelements) << "% "
                  << "Inserting element " << i << " top() == " << least.key << " @ "
                  << std::setprecision(3) << Timer.seconds() << " s"
                  << std::setprecision(6) << std::resetiosflags(std::ios_base::floatfield);
@@ -366,7 +366,7 @@ int main(int argc, char* argv[])
 
         if ((i % mega) == 0)
             LOG1 << std::fixed << std::setprecision(2) << std::setw(5)
-                 << (100.0 * (double)i / (double)nelements) << "% "
+                 << (100.0 * static_cast<double>(i) / nelements) << "% "
                  << "Popped element " << i << " == " << least.key << " @ "
                  << std::setprecision(3) << Timer.seconds() << " s"
                  << std::setprecision(6) << std::resetiosflags(std::ios_base::floatfield);
